feat(persistence): basic persistence works
