+++ src/main.rs
@@ -10,7 +10,7 @@ use indicatif::{
     HumanBytes, HumanDuration, ProgressBar, ProgressDrawTarget, ProgressState, ProgressStyle,
 };
 use iroh::protocol::AuthToken;
-use iroh::provider::{Database, Provider, Ticket};
+use iroh::provider::{Database, Provider, Snapshot, Ticket};
 use iroh::rpc_protocol::{
     ListRequest, ProvideRequest, ProvideResponse, ProvideResponseEntry, ProviderRequest,
     ProviderResponse, ProviderService, VersionRequest,
@@ -18,8 +18,10 @@ use iroh::rpc_protocol::{
 use quic_rpc::transport::quinn::{QuinnConnection, QuinnServerEndpoint};
 use quic_rpc::{RpcClient, ServiceEndpoint};
 use tracing_subscriber::{prelude::*, EnvFilter};
+mod main_util;
 
 use iroh::{get, provider, Hash, Keypair, PeerId};
+use main_util::{iroh_data_root, Blake3Cid};
 
 const DEFAULT_RPC_PORT: u16 = 0x1337;
 const RPC_ALPN: [u8; 17] = *b"n0/provider-rpc/1";
@@ -156,91 +158,6 @@ macro_rules! progress {
     }};
 }
 
-#[repr(transparent)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct Blake3Cid(Hash);
-
-const CID_PREFIX: [u8; 4] = [
-    0x01, // version
-    0x55, // raw codec
-    0x1e, // hash function, blake3
-    0x20, // hash size, 32 bytes
-];
-
-impl Blake3Cid {
-    pub fn new(hash: Hash) -> Self {
-        Blake3Cid(hash)
-    }
-
-    pub fn as_hash(&self) -> &Hash {
-        &self.0
-    }
-
-    pub fn as_bytes(&self) -> [u8; 36] {
-        let hash: [u8; 32] = self.0.as_ref().try_into().unwrap();
-        let mut res = [0u8; 36];
-        res[0..4].copy_from_slice(&CID_PREFIX);
-        res[4..36].copy_from_slice(&hash);
-        res
-    }
-
-    pub fn from_bytes(bytes: &[u8]) -> anyhow::Result<Self> {
-        anyhow::ensure!(
-            bytes.len() == 36,
-            "invalid cid length, expected 36, got {}",
-            bytes.len()
-        );
-        anyhow::ensure!(bytes[0..4] == CID_PREFIX, "invalid cid prefix");
-        let mut hash = [0u8; 32];
-        hash.copy_from_slice(&bytes[4..36]);
-        Ok(Blake3Cid(Hash::from(hash)))
-    }
-}
-
-impl fmt::Display for Blake3Cid {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // result will be 58 bytes plus prefix
-        let mut res = [b'b'; 59];
-        // write the encoded bytes
-        data_encoding::BASE32_NOPAD.encode_mut(&self.as_bytes(), &mut res[1..]);
-        // convert to string, this is guaranteed to succeed
-        let t = std::str::from_utf8_mut(res.as_mut()).unwrap();
-        // hack since data_encoding doesn't have BASE32LOWER_NOPAD as a const
-        t.make_ascii_lowercase();
-        // write the str, no allocations
-        f.write_str(t)
-    }
-}
-
-impl FromStr for Blake3Cid {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let sb = s.as_bytes();
-        if sb.len() == 59 && sb[0] == b'b' {
-            // this is a base32 encoded cid, we can decode it directly
-            let mut t = [0u8; 58];
-            t.copy_from_slice(&sb[1..]);
-            // hack since data_encoding doesn't have BASE32LOWER_NOPAD as a const
-            std::str::from_utf8_mut(t.as_mut())
-                .unwrap()
-                .make_ascii_uppercase();
-            // decode the bytes
-            let mut res = [0u8; 36];
-            data_encoding::BASE32_NOPAD
-                .decode_mut(&t, &mut res)
-                .map_err(|_e| anyhow::anyhow!("invalid base32"))?;
-            // convert to cid, this will check the prefix
-            Self::from_bytes(&res)
-        } else {
-            // if we want to support all the weird multibase prefixes, we have no choice
-            // but to use the multibase crate
-            let (_base, bytes) = multibase::decode(s)?;
-            Self::from_bytes(bytes.as_ref())
-        }
-    }
-}
-
 async fn make_rpc_client(
     rpc_port: u16,
 ) -> anyhow::Result<RpcClient<ProviderService, QuinnConnection<ProviderResponse, ProviderRequest>>>
@@ -352,7 +269,13 @@ async fn main_impl() -> Result<()> {
             key,
             rpc_port,
         } => {
-            let provider = provide(addr, auth_token, key, cli.keylog, rpc_port.into()).await?;
+            let iroh_data_root = iroh_data_root()?;
+            tracing::info!("Loading snapshot from {}...", iroh_data_root.display());
+            let snapshot = Snapshot::load(&iroh_data_root)?;
+            tracing::info!("Loading database from snapshot");
+            let db = Database::from_snapshot(snapshot)?;
+
+            let provider = provide(db, addr, auth_token, key, cli.keylog, rpc_port.into()).await?;
             let controller = provider.controller();
             let mut ticket = provider.ticket(Hash::from([0u8; 32]));
 
@@ -385,7 +308,11 @@ async fn main_impl() -> Result<()> {
             tokio::signal::ctrl_c().await?;
             println!("Shutting down provider...");
             provider.shutdown();
+            let database = provider.database().clone();
             provider.await?;
+            let snapshot = database.snapshot();
+            println!("Saving snapshot to {}...", iroh_data_root.display());
+            snapshot.persist(iroh_data_root)?;
             // the future holds a reference to the temp file, so we need to
             // keep it for as long as the provider is running. The drop(fut)
             // makes this explicit.
@@ -419,6 +346,7 @@ async fn main_impl() -> Result<()> {
 }
 
 async fn provide(
+    db: Database,
     addr: Option<SocketAddr>,
     auth_token: Option<String>,
     key: Option<PathBuf>,
@@ -427,7 +355,6 @@ async fn provide(
 ) -> Result<Provider> {
     let keypair = get_keypair(key).await?;
 
-    let db = Database::default();
     let mut builder = provider::Provider::builder(db).keylog(keylog);
     if let Some(addr) = addr {
         builder = builder.bind_addr(addr);
