+++ src/util.rs
@@ -1,19 +1,15 @@
+//! Utility functions and types.
 use std::{
-    env,
     fmt::{self, Display},
-    path::{Path, PathBuf},
     result,
     str::FromStr,
 };
 
-use anyhow::{anyhow, ensure, Result};
+use anyhow::{ensure, Result};
 use base64::{engine::general_purpose, Engine as _};
 use postcard::experimental::max_size::MaxSize;
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use thiserror::Error;
-
-/// name of directory that wraps all iroh files in a given application directory
-const IROH_DIR: &str = "iroh";
 
 /// Encode the given buffer into Base64 URL SAFE without padding.
 pub fn encode(buf: impl AsRef<[u8]>) -> String {
@@ -153,93 +149,8 @@ impl From<anyhow::Error> for RpcError {
     }
 }
 
+/// A serializable result type for use in RPC responses.
 pub type RpcResult<T> = result::Result<T, RpcError>;
-
-/// Returns the path to the user's iroh config directory.
-///
-/// If the `IROH_CONFIG_DIR` environment variable is set it will be used unconditionally.
-/// Otherwise the returned value depends on the operating system according to the following
-/// table.
-///
-/// | Platform | Value                                 | Example                          |
-/// | -------- | ------------------------------------- | -------------------------------- |
-/// | Linux    | `$XDG_CONFIG_HOME` or `$HOME`/.config/iroh | /home/alice/.config/iroh              |
-/// | macOS    | `$HOME`/Library/Application Support/iroh   | /Users/Alice/Library/Application Support/iroh |
-/// | Windows  | `{FOLDERID_RoamingAppData}`/iroh           | C:\Users\Alice\AppData\Roaming\iroh   |
-#[allow(dead_code)]
-pub fn iroh_config_root() -> Result<PathBuf> {
-    if let Some(val) = env::var_os("IROH_CONFIG_DIR") {
-        return Ok(PathBuf::from(val));
-    }
-    let cfg = dirs_next::config_dir()
-        .ok_or_else(|| anyhow!("operating environment provides no directory for configuration"))?;
-    Ok(cfg.join(IROH_DIR))
-}
-
-// Path that leads to a file in the iroh config directory.
-#[allow(dead_code)]
-pub fn iroh_config_path(file_name: impl AsRef<Path>) -> Result<PathBuf> {
-    let path = iroh_config_root()?.join(file_name);
-    Ok(path)
-}
-
-/// Returns the path to the user's iroh data directory.
-///
-/// If the `IROH_DATA_DIR` environment variable is set it will be used unconditionally.
-/// Otherwise the returned value depends on the operating system according to the following
-/// table.
-///
-/// | Platform | Value                                         | Example                                  |
-/// | -------- | --------------------------------------------- | ---------------------------------------- |
-/// | Linux    | `$XDG_DATA_HOME`/iroh or `$HOME`/.local/share/iroh | /home/alice/.local/share/iroh                 |
-/// | macOS    | `$HOME`/Library/Application Support/iroh      | /Users/Alice/Library/Application Support/iroh |
-/// | Windows  | `{FOLDERID_RoamingAppData}/iroh`              | C:\Users\Alice\AppData\Roaming\iroh           |
-#[allow(dead_code)]
-pub fn iroh_data_root() -> Result<PathBuf> {
-    if let Some(val) = env::var_os("IROH_DATA_DIR") {
-        return Ok(PathBuf::from(val));
-    }
-    let path = dirs_next::data_dir().ok_or_else(|| {
-        anyhow!("operating environment provides no directory for application data")
-    })?;
-    Ok(path.join(IROH_DIR))
-}
-
-/// Path that leads to a file in the iroh data directory.
-#[allow(dead_code)]
-pub fn iroh_data_path(file_name: impl AsRef<Path>) -> Result<PathBuf> {
-    let path = iroh_data_root()?.join(file_name);
-    Ok(path)
-}
-
-/// Returns the path to the user's iroh cache directory.
-///
-/// If the `IROH_CACHE_DIR` environment variable is set it will be used unconditionally.
-/// Otherwise the returned value depends on the operating system according to the following
-/// table.
-///
-/// | Platform | Value                                         | Example                                  |
-/// | -------- | --------------------------------------------- | ---------------------------------------- |
-/// | Linux    | `$XDG_CACHE_HOME`/iroh or `$HOME`/.cache/iroh | /home/.cache/iroh                        |
-/// | macOS    | `$HOME`/Library/Caches/iroh                   | /Users/Alice/Library/Caches/iroh         |
-/// | Windows  | `{FOLDERID_LocalAppData}/iroh`                | C:\Users\Alice\AppData\Roaming\iroh      |
-#[allow(dead_code)]
-pub fn iroh_cache_root() -> Result<PathBuf> {
-    if let Some(val) = env::var_os("IROH_CACHE_DIR") {
-        return Ok(PathBuf::from(val));
-    }
-    let path = dirs_next::cache_dir().ok_or_else(|| {
-        anyhow!("operating environment provides no directory for application data")
-    })?;
-    Ok(path.join(IROH_DIR))
-}
-
-/// Path that leads to a file in the iroh cache directory.
-#[allow(dead_code)]
-pub fn iroh_cache_path(file_name: impl AsRef<Path>) -> Result<PathBuf> {
-    let path = iroh_cache_root()?.join(file_name);
-    Ok(path)
-}
 
 #[cfg(test)]
 mod tests {
